+++ src/restaurantManagementModule/FastFoodRestaurant.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 public class FastFoodRestaurant extends Restaurant {
 
 
-    public FastFoodRestaurant(long id, String name, String location, LocalTime closeTime, LocalTime openTime, boolean status) {
+    public FastFoodRestaurant(int id, String name, String location, LocalTime closeTime, LocalTime openTime, boolean status) {
         super(id, name, CuisineType.FAST_FOOD_RESTAURANT, location, closeTime, openTime, status);
     }
