+++ src/restaurantManagementModule/FineDiningRestaurant.java
@@ -4,7 +4,7 @@ import java.time.LocalTime;
 
 public class FineDiningRestaurant extends Restaurant {
 
-    public FineDiningRestaurant(long id, String name, String location, LocalTime closeTime, LocalTime openTime, boolean status) {
+    public FineDiningRestaurant(int id, String name, String location, LocalTime closeTime, LocalTime openTime, boolean status) {
         super(id, name, CuisineType.FINE_DINING_RESTAURANT, location, closeTime, openTime, status);
     }
